Add Bare Metal SIG to governance
The purpose of the Bare Metal SIG will be to promote the development and
use of Ironic and other OpenStack bare-metal software. This will include
marketing efforts like case studies of Ironic clusters in industry and
academia, supporting integration of Ironic with projects like Airship and
the Kubernetes Cluster API, coordinating presentations for industry
events, developing documentation and tutorials, gathering feedback from
the community on usage and feature gaps, and other broader
community-facing efforts to encourage the adoption of Ironic as a
bare-metal management tool.

http://lists.openstack.org/pipermail/openstack-discuss/2019-January/002177.html
https://etherpad.openstack.org/p/bare-metal-sig

Change-Id: I7e84bb66a2e63f560e96f2d5c79d4cf3a1de9db9