Update SIG admins
According to governance[1] approved by both the TC and the
UC, SIGs are administered by a UC representative and a TC
representative (also serving as Meta SIG co-chairs).

Following the recent leadership elections, Matt Van Winkle
has volunteered for the UC[2] and Rico Lin has volunteered
for the TC[3].

Upon approval of this patch, I'll update the sig-admins Gerrit
group so that the Meta SIG co-chairs can approve future changes
to the openstack/governance-sigs repository.

[1] https://governance.openstack.org/tc/resolutions/20180319-sig-governance.html
[2] http://eavesdrop.openstack.org/meetings/uc/2019/uc.2019-03-04-16.00.log.html#l-45
[3] http://eavesdrop.openstack.org/irclogs/%23openstack-tc/%23openstack-tc.2019-03-07.log.html#t2019-03-07T15:06:30

Change-Id: I99995ca01e63364fa8e31f153ddf4e82f9ec0b97