Propose new Hardware Vendor SIG
This proposes the creation of a new SIG for those working on integration
with different hardware platforms. This is to enable collaboration and
open development for things like drivers, tooling, libraries, and any
other needs related to specific hardware support. The hope is this will
allow vendors to work together and share code and expertise to improve
overall hardware support.

Change-Id: If7f47e1cb12b27f5118dcd2d16464fac10463534
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>