The Meta SIG is no longer needed
The Meta SIG was created to administrate and promote the SIGs.
This was needed as the SIGs were co-owned by the OpenStack TC
and the OpenStack UC. Now that they operate as a single group,
we can just put the SIGs under the responsibility of the TC,
like other community work groups in OpenStack.

Change-Id: I1f6e3aa90d3f83c7d37df4e726f3e983ae948bff