Create the Packaging SIG
Nowadays, the rpm-packaging team can be moved to a lightweight
governance model that the SIG brings. We should move the
rpm-packaging to a SIG to make it easier for the contributors,
they wouldn't need to worry about elections in the future either,
just giving a regular heartbeat would be enough.

Change-Id: I4cafe9bb05c9e50d24286c4d6add0705338949d9