Add doc/requirements
This project isn't released but uniformization can't hurt.

The origin of these changes was that we need to specify doc requirements
in doc/requirements.txt to avoid problems with the pip resolver [1]
for the release team [2][3].

The problem here is that this repos haven't doc/requirements.txt
file and by default in this case zuul will use the test-requirements.txt
file to pull requirements [4].

This requirements file contains extra requirements like flake8 that
collided with those allowed in our job environment and so the new
pip resolver fails to install these requirements and the job exits in
error.

This project could meet similar conditions leading to the bug.

[1] http://lists.openstack.org/pipermail/release-job-failures/2021-January/001500.html
[2] http://lists.openstack.org/pipermail/openstack-discuss/2021-January/019611.html
[3] http://lists.openstack.org/pipermail/openstack-discuss/2021-January/019612.html
[4] https://opendev.org/zuul/zuul-jobs/src/branch/master/roles/ensure-sphinx/tasks/main.yaml#L36

Change-Id: I0101b340fef5ae465eb64e7d7b21973d36fd2a6a