Mark 'Technical Writing' SIG complete as it is merged into TC
The Technical Writing SIG has served its purpose to provde
the guidance and tooling to have most of the documentation
ownership to projects. Now we do not have much work in this
SIG and it is also lacking maintainer to lead this SIG. The
current Chair Stephen Finucane would not be able to continue
leading this SIG.

In TC weekly meetings and PTG[1], we decided to merge this
SIG responsibility in TC (depends-on). SIG repos will be moved
to TC and a few repos which are related to upstream institute training
will be moved to First Contact SIG.

One repo 'training-labs' is left for now, which will be called
for the maintainer, and if there is no one, we can retire it.
We are not even sure who is using it. Upstream training and CoA
do not use this repo.

Depends-On: https://review.opendev.org/c/openstack/governance/+/815869

[1] https://etherpad.opendev.org/p/tc-yoga-ptg
    http://lists.openstack.org/pipermail/openstack-discuss/2021-October/025554.html

Change-Id: I5f2061e788346348b4f3ec375e9498a368c4ac47